Optimize memory usage.
This change reduces the memory usage per slot and makes further optimization easier, and extends the error handling strategy to handle more edge cases.

This change is an alternative to the reverted change that tolerated unaligned memory access, which is at the hardware-level (Intel and ARM) just fine, but it could be a very bad idea in Rust (I'm told), because of its current optimization patterns. So the taken advice is that unaligned memory access in Rust is never ok.

This change also expands the "willy-nilly" error handling strategy to guard allocations without creating a total mess. Since this ship has sailed, error handling can't remain a "willy-nilly" thing.

This commit is relatively large, but it was much easier than multiple staged commits with intermediate hacks.